Cambiar float por numeric en migraciones y agregar campo importe en ventasCabeceraComponent

+++ database/migrations/2019_08_12_211052_create_articulos_table.php
@@ -17,8 +17,8 @@ class CreateArticulosTable extends Migration
             $table->bigIncrements('id');
             $table->bigInteger('tipo_id')->unsigned();
             $table->string('nombre');
-            $table->float('precio', 10, 2);
-            $table->float('costo', 10, 2);
+            $table->decimal('precio', 10, 2);
+            $table->decimal('costo', 10, 2);
 
             $table->foreign('tipo_id')->references('id')->on('tipos');
         });
