+++ database/migrations/2019_08_12_211322_create_articulos_insumos_table.php
@@ -17,7 +17,7 @@ class CreateArticulosInsumosTable extends Migration
             $table->bigIncrements('id');
             $table->bigInteger('articulo_id')->unsigned();
             $table->bigInteger('insumo_id')->unsigned();
-            $table->float('cantidad', 10, 2);
+            $table->decimal('cantidad', 10, 2);
 
             $table->foreign('articulo_id')->references('id')->on('articulos');
             $table->foreign('insumo_id')->references('id')->on('insumos');
